// frontend/src/openapi/models/TraceGroup.ts
export interface TraceGroup {
    /**
     *
     * @type {string}
     * @memberof TraceGroup
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof TraceGroup
     */
    name: string;
    /**
     *
     * @type {number}
     * @memberof TraceGroup
     */
    revision: number;
    /**
     *
     * @type {number}
     * @memberof TraceGroup
     */
    traceCount: number;
    /**
     *
     * @type {number}
     * @memberof TraceGroup
     */
    unreviewedTraceCount: number;
    /**
     *
     * @type {boolean}
     * @memberof TraceGroup
     */
    needsReview: boolean;
    /**
     *
     * @type {string}
     * @memberof TraceGroup
     */
    capabilityId: string | null;
    /**
     *
     * @type {boolean}
     * @memberof TraceGroup
     */
    mixed: boolean;
    /**
     *
     * @type {TraceGroupEvidence}
     * @memberof TraceGroup
     */
    evidence: TraceGroupEvidence;
    /**
     *
     * @type {Array<string>}
     * @memberof TraceGroup
     */
    sampleTraceIds: Array<string>;
    /**
     *
     * @type {Date}
     * @memberof TraceGroup
     */
    lastReviewedAt: Date | null;
    /**
     *
     * @type {number}
     * @memberof TraceGroup
     */
    lastReviewedTraceCount: number | null;
}